+++ shoop/testing/models/__init__.py
@@ -1,0 +1,16 @@
+# This file is part of Shoop.
+#
+# Copyright (c) 2012-2016, Shoop Ltd. All rights reserved.
+#
+# This source code is licensed under the AGPLv3 license found in the
+# LICENSE file in the root directory of this source tree.
+
+from ._behavior_components import ExpensiveSwedenBehaviorComponent
+from ._payment_with_phase import PaymentWithCheckoutPhase
+from ._pseudo_payment import PseudoPaymentProcessor
+
+__all__ = [
+    "ExpensiveSwedenBehaviorComponent",
+    "PaymentWithCheckoutPhase",
+    "PseudoPaymentProcessor",
+]