+++ shoop_tests/campaigns/__init__.py
@@ -6,7 +6,7 @@
 # LICENSE file in the root directory of this source tree.
 from django.utils.translation import activate
 from shoop.testing.factories import (
-    get_shop, get_default_customer_group, get_default_payment_method,
+    get_shop, get_default_customer_group, get_payment_method,
     create_random_person
 )
 from shoop.testing.utils import apply_request_middleware
@@ -15,7 +15,11 @@ from shoop.testing.utils import apply_request_middleware
 def initialize_test(rf, include_tax=False):
     activate("en")
     shop = get_shop(prices_include_tax=include_tax)
-    get_default_payment_method()  # Valid baskets needs some payment methods to be available
+
+    # Valid baskets needs some payment methods to be available
+    get_payment_method(shop)
+    # Since some of the baskets are created for the default shop:
+    get_payment_method(None)
 
     group = get_default_customer_group()
     customer = create_random_person()