+++ doc/changelog.rst
@@ -12,6 +12,8 @@ Unreleased
 Core
 ~~~~
 
+- Add new Service API and implement shipping and payments with it
+- Remove BaseMethodModule based API (``shoop.core.methods``)
 - Add management command to generate bought with relations
 - Add ``merchant_notes`` text-field to ``Contact``
 - Add ``first_name`` and ``last_name`` fields to ``PersonContact``
@@ -47,6 +49,7 @@ Localization
 Admin
 ~~~~~
 
+- Remove shipping and payment method management
 - Add package mode for products
 - Enable ```merchant_notes`` editing for ``Contact``
 - Add option to add extra form parts to Shop edit view
@@ -60,6 +63,7 @@ Admin
 Front
 ~~~~~
 
+- Remove ``get_method_validation_errors`` signal
 - Fix bug at ``get_visible_products`` filter when orderable_only is False
 - Set template price display options from the customer
 - Fix bug: BasketStorage.finalize() never called delete() correctly